[FIX] fix delete team bug

+++ console/views/team.py
@@ -522,6 +522,11 @@ class TeamDelView(JWTAuthApiView):
                 if service_count >= 1:
                     result = general_message(400, "failed", "当前团队内有应用,不可以删除")
                     return Response(result, status=400)
+
+                tenants = team_services.get_current_user_tenants(self.user.user_id)
+                if len(tenants) == 1:
+                    return Response(general_message(409, "you have to keep one team", "您必须保留一个团队"), status=409)
+
                 status = team_services.delete_tenant(tenant_name=team_name)
                 if not status:
                     result = general_message(code, "delete a tenant successfully", "删除团队成功")
